Re-organize DailyNoteEditor
Use renderResult in DailyNoteEditor

Tweak daily note editor layout according to the new design

Remove redundant default fallback in Title component

Tweak group note editor layout according to the new design

Fix visually broken clear buttons in note editors

Remove redundant containers
